// src/app/models/notice.model.ts
export class Notice {
  sort(arg0: (a: any, b: any) => 1 | 0 | -1) {
    throw new Error('Method not implemented.');
  }
      id: any;
	  posted_by: any;
	  user_photo: any;
	  posted_at: any;
	  notice_details: any;
	  notice_type: any;
	  notice_photo: any;
	  user_email:any;
	  notice_likes:any;
	  userphoneNumber:any;
}



// export class Notice {
//     id: number = 0;
//     posted_by: string = '';
//     user_photo: string = '';
//     posted_at: string = '';
//     notice_details: string = '';
//     notice_type: string = '';
// }